+++ dumux/material/fluidmatrixinteractions/frictionlaws/nofriction.hh
@@ -43,17 +43,18 @@ public:
     FrictionLawNoFriction() = default;
 
     /*!
-     * \brief Compute the shear stress.
+     * \brief Compute the bottom shear stress.
      *
      * \param volVars Volume variables
      *
-     * Compute the shear stress due to friction. The shear stress is not a tensor as know
-     * from contiuums mechanics, but a force projected on an area. Therefore it is a
-     * vector with two entries. For this law without friction, the shearStress is zero.
+     * Compute the bottom shear stress due to bottom friction.
+     * The bottom shear stress is a projection of the shear stress tensor onto the river bed.
+     * It can therefore be represented by a (tangent) vector with two entries.
+     * For this law without bottom friction, the bottom shear stress is zero.
      *
-     * \return shear stress [N/m^2]. First entry is the x-component, the second the y-component.
+     * \return shear stress in N/m^2. First entry is the x-component, the second the y-component.
      */
-    Dune::FieldVector<Scalar, 2> shearStress(const VolumeVariables& volVars) const final
+    Dune::FieldVector<Scalar, 2> bottomShearStress(const VolumeVariables& volVars) const final
     {
         return {0.0, 0.0};
     }